fixed spring-cloud context integration tests - part 2:
