+++ spring-cloud/spring-cloud-config/client/src/test/java/com/baeldung/spring/cloud/config/client/SpringContextLiveTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.spring.cloud.config.server;
+package com.baeldung.spring.cloud.config.client;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -6,17 +6,16 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 
-
 /**
  * 
- * The context will load successfully with some properties provided by docker
+ * The app needs the server running on port 8888. Can be started with docker
  *
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringBootTest(classes = ConfigServer.class)
+@SpringBootTest(classes = ConfigClient.class)
 @WebAppConfiguration
 public class SpringContextLiveTest {
     @Test
-    public void whenSpringContextIsBootstrapped_thenNoExceptions() {
+    public void contextLoads() {
     }
 }
