+++ spring-cloud/spring-cloud-config/server/src/test/java/com/baeldung/spring/cloud/config/server/SpringContextIntegrationTest.java
@@ -1,4 +1,4 @@
-package org.baeldung;
+package com.baeldung.spring.cloud.config.server;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -6,13 +6,11 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 
-import com.baeldung.spring.cloud.config.client.ConfigClient;
-
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringBootTest(classes = ConfigClient.class)
+@SpringBootTest(classes = ConfigServer.class)
 @WebAppConfiguration
 public class SpringContextIntegrationTest {
     @Test
-    public void contextLoads() {
+    public void whenSpringContextIsBootstrapped_thenNoExceptions() {
     }
 }
