+++ spring-cloud/spring-cloud-config/server/src/test/resources/application.properties
@@ -1,0 +1,2 @@
+### This should be provided by the docker config
+spring.cloud.config.server.git.uri=classpath:.
