+++ spring-cloud/spring-cloud-connectors-heroku/src/main/resources/application.properties
@@ -4,5 +4,6 @@ spring.datasource.maxIdle=5
 spring.datasource.minIdle=2
 spring.datasource.initialSize=5
 spring.datasource.removeAbandoned=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 
 spring.jpa.hibernate.ddl-auto=update
