+++ spring-cloud/spring-cloud-connectors-heroku/src/test/resources/application.properties
@@ -1,0 +1,2 @@
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
