+++ spring-cloud/spring-cloud-consul/src/test/java/com/baeldung/spring/cloud/consul/SpringContextLiveTest.java
@@ -1,4 +1,4 @@
-package org.baeldung;
+package com.baeldung.spring.cloud.consul;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -9,10 +9,18 @@ import com.baeldung.spring.cloud.consul.discovery.DiscoveryClientApplication;
 import com.baeldung.spring.cloud.consul.health.ServiceDiscoveryApplication;
 import com.baeldung.spring.cloud.consul.properties.DistributedPropertiesApplication;
 
+
+/**
+ * 
+ * This Live test requires:
+ * * a Consul instance running on port 8500
+ * * Consul configured with particular properties (e.g. 'my.prop')
+ *
+ */
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = { DiscoveryClientApplication.class, ServiceDiscoveryApplication.class,
 		DistributedPropertiesApplication.class })
-public class SpringContextIntegrationTest {
+public class SpringContextLiveTest {
 
     @Test
     public void whenSpringContextIsBootstrapped_thenNoExceptions() {
